Remove DRep voting workflows and associated scripts to streamline the codebase, eliminating unused files and simplifying the governance process.
