--- .github/workflows/mesh-gov-update-rationales.yml
@@ -1,48 +1,0 @@
-name: Update Voting Rationales
-
-on:
-  #schedule:
-  #  - cron: '0 4 * * *'  # Runs at 4 AM UTC every day
-  #  - cron: '0 16 * * *'  # Runs at 4 PM UTC every day
-  workflow_dispatch:
-
-jobs:
-  update-rationales:
-    runs-on: ubuntu-latest
-    
-    steps:
-    - uses: actions/checkout@v4
-    
-    - name: Setup Node.js
-      uses: actions/setup-node@v4
-      with:
-        node-version: '20'
-        
-    - name: Install dependencies
-      run: npm install
-      
-    - name: Run update rationales script
-      env:
-        KOIOS_API_KEY: ${{ secrets.KOIOS_API_KEY }}
-      run: node scripts/drep-voting/update-rationales.js
-
-    - name: Check for changes
-      id: git-check
-      run: |
-        git status --porcelain
-        echo "changes=$(git status --porcelain | wc -l)" >> $GITHUB_OUTPUT
-
-    - name: Commit files
-      if: steps.git-check.outputs.changes > 0
-      run: |
-        git config --local user.email "github-actions[bot]@users.noreply.github.com"
-        git config --local user.name "github-actions[bot]"
-        git add voting-history/missing-voting-rationales/rationales.json
-        git commit -m "Update voting rationales"
-        
-    - name: Push changes
-      if: steps.git-check.outputs.changes > 0
-      uses: ad-m/github-push-action@master
-      with:
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-        branch: ${{ github.ref }}
